153: Refactor WS module
